+++ src/bittytax/conv/parsers/revolut.py
@@ -81,6 +81,17 @@ def parse_revolut(data_row: "DataRow", parser: DataParser, **_kwargs: Unpack[Par
             buy_asset=row_dict["Symbol"],
             wallet=WALLET,
         )
+    elif row_dict["Type"] == "Staking reward":
+        data_row.t_record = TransactionOutRecord(
+            TrType.STAKING_REWARD,
+            data_row.timestamp,
+            buy_quantity=Decimal(row_dict["Quantity"].replace(",", "")),
+            buy_asset=row_dict["Symbol"],
+            wallet=WALLET,
+        )
+    elif row_dict["Type"] in ("Stake", "Unstake"):
+        # Skip internal transfers
+        return
     else:
         raise UnexpectedTypeError(parser.in_header.index("Type"), "Type", row_dict["Type"])